requirements.txt
Cython==0.23.4
falcon==0.3.0
Janome==0.2.6
pykintone==0.2.0.1
pymongo==3.1.1
python-mimeparse==0.1.4
PyYAML==3.11
requests==2.8.1
six==1.10.0
waitress==0.8.10
wheel==0.24.0
